# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema POS Ventas API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brands:
    get:
      operationId: getBrands
      tags: [brands]
      summary: List all brands
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [brands]
      summary: Create brand
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "201":
          description: Brand created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    put:
      operationId: updateBrand
      tags: [brands]
      summary: Update brand
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "200":
          description: Brand updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
    delete:
      operationId: deleteBrand
      tags: [brands]
      summary: Delete brand
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brand deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # PRODUCTS
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with optional search
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: brandId
          in: query
          required: false
          schema:
            type: integer
        - name: lowStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProductById
      tags: [products]
      summary: Get product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/inventory-report:
    get:
      operationId: getInventoryReport
      tags: [products]
      summary: Get full inventory data for Excel export
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Inventory report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryReportItem"

  # CUSTOMERS
...
  /business-settings:
    get:
      operationId: getBusinessSettings
      tags: [business]
      summary: Get business configuration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Business settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"
    put:
      operationId: updateBusinessSettings
      tags: [business]
      summary: Update business configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"

  # PAYMENT METHODS
  /payment-methods:
    get:
      operationId: getPaymentMethods
      tags: [payment-methods]
      summary: List payment methods
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
    post:
      operationId: createPaymentMethod
      tags: [payment-methods]
      summary: Create payment method
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentMethodBody"
      responses:
        "201":
          description: Payment method created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
  /payment-methods/{id}:
    put:
      operationId: updatePaymentMethod
      tags: [payment-methods]
      summary: Update payment method
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentMethodBody"
      responses:
        "200":
          description: Payment method updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
    delete:
      operationId: deletePaymentMethod
      tags: [payment-methods]
      summary: Delete payment method
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USERS MANAGEMENT
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    Brand:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        productCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, productCount, createdAt]
    CreateBrandBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required: [name]

    # PRODUCTS
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        barcode:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        purchasePrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        minStock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        brandId:
          type: ["integer", "null"]
        brandName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, purchasePrice, salePrice, stock, minStock, createdAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        barcode:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        purchasePrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        minStock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
      required: [name, purchasePrice, salePrice, stock, minStock]
...
    InventoryReportItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        barcode:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        brandName:
          type: ["string", "null"]
        purchasePrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        minStock:
          type: integer
        stockValue:
          type: number
        status:
          type: string
        imageUrl:
          type: ["string", "null"]
      required: [id, name, purchasePrice, salePrice, stock, minStock, stockValue, status]

    # BUSINESS SETTINGS
    BusinessSettings:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        rucNit:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        currency:
          type: string
        currencySymbol:
          type: string
        updatedAt:
          type: string
      required: [id, companyName, currency, currencySymbol, updatedAt]
    UpdateBusinessSettingsBody:
      type: object
      properties:
        companyName:
          type: string
        rucNit:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        currency:
          type: string
        currencySymbol:
          type: string
      required: [companyName, currency, currencySymbol]

    # PAYMENT METHODS
    PaymentMethod:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, isActive, createdAt]
    CreatePaymentMethodBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [name]

    # REPORTS
...